refactor event tile into a reusable stateless widget class
